complete seminar 01; start seminar 02; fix some .md

+++ 2019_hse_cpp_seminars/remove_trash.py
@@ -9,7 +9,7 @@ def main():
     args = parser.parse_args()
 
     for path in os.listdir(args.dir):
-        if not path.endswith('.cpp'):
+        if not path.endswith('.cpp') and not path.endswith('.py'):
             os.remove(os.path.join(args.dir, path))
